[模糊搜索]

+++ java/api/src/main/java/com/exadel/frs/core/trainservice/exel/VideoExelRow.java
@@ -20,7 +20,7 @@ public class VideoExelRow
 
     private String  deviceIdAddress;
 
-    private Integer  timestamp;
+    private long  timestamp;
 
     private String  VideImg;
